requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
langchain-anthropic==0.1.1
chromadb==0.4.22
pymongo==4.6.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyyaml==6.0.1
sentence-transformers==2.3.1
faiss-cpu==1.8.0
tiktoken==0.5.2
numpy==1.26.3
